pyproject.toml
[project]
name = "kottapalli-archive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "indic-transliteration>=2.3.79",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
]
